+++ package.json
@@ -6,7 +6,6 @@
   "license": "MIT",
   "dependencies": {
     "@cosmostation/cosmosjs": "^0.5.2",
-    "@material-ui/core": "^4.9.3",
     "@types/base-64": "^0.1.3",
     "@types/react-image-crop": "^3.0.0",
     "@types/react-redux": "^7.1.5",
@@ -38,10 +37,8 @@
     "react-dom": "^16.12.0",
     "react-dropzone": "^4.2.13",
     "react-ga": "^2.7.0",
-    "react-grid-dropdown": "^0.3.1",
     "react-hook-form": "^3.29.4",
     "react-image-crop": "^4.0.1",
-    "react-laag": "^1.7.2",
     "react-mailchimp-subscribe": "^2.1.0",
     "react-md-file": "^1.1.0",
     "react-modal": "^3.11.1",
