--- src/assets/filter_sort_buttons.css
@@ -1,35 +1,0 @@
-.btn1 {
-  background-color: white;
-  border-color: grey;
-  border-width: 0.5px;
-  margin: 8px;
-  font-weight: 500;
-  font-style: normal;
-  font-family: Roboto;
-  font-size: 16px;
-  padding: 4px 6px;
-  border-radius: 4px;
-  align-items: center;
-  line-height: 19px;
-  height: 40px;
-  width: 100px;
-}
-
-.btn1-focus {
-  background-color: white;
-  color: black;
-}
-
-.btn1-focus:hover {
-  border-color: grey;
-  border-width: 2px;
-  color: black;
-}
-
-.icon-calendar-sort {
-  padding-right: 10px;
-}
-
-.icon-reset {
-  padding-right: 10px;
-}
