+++ src/lib/json/filterData.json
@@ -4,7 +4,7 @@
   "categories": [
     {
       "@type": "Category",
-      "@id": "{some id}",
+      "@id": "1",
       "title": "Geo Area",
       "tags": [
         {
@@ -35,7 +35,7 @@
     },
     {
       "@type": "Category",
-      "@id": "{some id}",
+      "@id": "2",
       "title": "Goal",
       "tags": [
         {
@@ -66,7 +66,7 @@
     },
     {
       "@type": "Category",
-      "@id": "{some id}",
+      "@id": "3",
       "title": "Target",
       "tags": [
         {
@@ -97,7 +97,7 @@
     },
     {
       "@type": "Category",
-      "@id": "{some id}",
+      "@id": "4",
       "title": "Indicator",
       "tags": [
         {
